refactor prms_to_mf6 to mmr_to_mf6_mmr and separate files for class and its base mmr_to_mf6
